crawler/requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
lxml==4.9.3 